Refactor loading spinner into component
