+++ client/src/assets/stylesheets/Loader.css
@@ -1,0 +1,17 @@
+/* Loading spinner */
+.pizza-loader-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  background: #9FA4F5;
+}
+
+.pizza-loader-container p {
+  font-family: 'helveticaBold';
+  font-size: 24px;
+  color: var(--black);
+  margin-top: 15px;
+}
