+++ client/src/assets/stylesheets/RestaurantDetail.css
@@ -1,19 +1,4 @@
-.pizza-loader-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  background: #9FA4F5;
-}
-
-.pizza-loader-container p {
-  font-family: 'helveticaBold';
-  font-size: 24px;
-  color: var(--black);
-  margin-top: 15px;
-}
+@import 'variables.css';
 
 .restaurant-detail-wrapper {
   display: flex;
