+++ client/src/assets/stylesheets/RestaurantIndex.css
@@ -1,3 +1,5 @@
+@import 'variables.css';
+
 .new-restaurant-thumbnails-wrapper {
   width: 100%;
 }
