+++ client/src/assets/stylesheets/RestaurantMap.css
@@ -1,3 +1,5 @@
+@import 'variables.css';
+
 .full-map-container {
   position: relative;
   width: 100%;
@@ -27,7 +29,7 @@
   position: absolute;
   top: 230px;
   left: 250px;
-  padding: 15px 50px 15px 15px;
+  padding: 12px 50px 15px 15px;
   border: 2px solid var(--light-border-color);
   border-radius: 5px;
   box-sizing: border-box;
@@ -75,13 +77,3 @@
   color: var(--black);
   line-height: 1.15;
 }
-
-/* 
-#restaurant-map.single {
-    width: 100%;
-    height: 100%;
-}
-#restaurant-map.list {
-    width: 500px;
-    height: 500px;
-} */
